Calculate Growing Degree Days (GDD) from a Data Frame

The GDD which is a the growing degree days to the end date is calculated for each study year from a data frame. with date and temperature columns. date, which must be of class Date provides the dates and temperature which must be a numeric vector provides the mean daily water temperature in degrees centigrade.

Usage

gdd(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-09-30"),
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "longest",
  msgs = TRUE
)

Arguments

x

A data frame with two columns date and temperature.

start_date

A Date scalar of the first date within each year to consider (the year is ignored).

end_date

A Date scalar of the end date for the growing degree days (the year is ignored). If end_date is less than start_date (ignoring the year) then the window is considered to span two calendar years.

ignore_truncation

A flag specifying whether to ignore start truncation (end truncation is always ignored).

start_temp

A positive real number of the average water temperature at the start of the growing season in C.

end_temp

A positive real number of the average water temperature at the end of the growing season in C. It must be greater than or equal to the start temperature.

window_width

A positive whole number of the width of the rolling mean window in days. By default 7.

pick

A string specifying whether to pick the "longest", "shortest", "first" or "last" 'season' or the season with the "biggest" or "smallest" GSDD. By default the returned value is the the GSDD value for the "longest" 'season'.

msgs

A flag specifying whether to provide messages.

Value

A tibble with two columns year and gdd. year, which is an integer vector, indicates the year in which the window began and gdd which is a non-negative real number provides the GSDD or a missing value if it cannot be calculated.

See also

Examples

data <- gsdd::simulated_data
data$temperature <- data$synthetic
gdd(data)
#> # A tibble: 1 × 2
#>    year   gdd
#>   <int> <dbl>
#> 1  2019 3605.